Fix broken Buildkite build
Add .ruby-version
Disable the 'bundle install --deployment' option
Remove Bundler from the dependencies

As far as I understand:

- not having the .ruby-version prevented the gems to be installed
  in their expected location
- the --deployment option also prevented that, even when
  the .ruby-version file was present

Also, it seems that Bundler ignores itself when installing gems.
Since the 1.7.15 version was available (and in use) when installing
the bundle, the 1.11 version was never installed and was found
missing afterwards.